+++ utils/src/lib.rs
@@ -1,9 +1,8 @@
 //! A collection of utility functions and constants that can be reused from
 //! multiple projects
 
-// Enable unstable `is_multiple_of` on nightly for Wasm builds until nightly is updated
+// the feature `unsigned_is_multiple_of` has been stable since 1.87.0 and no longer requires an attribute to enable
 // See: https://github.com/o1-labs/mina-rust/issues/1997
-#![cfg_attr(target_arch = "wasm32", feature(unsigned_is_multiple_of))]
 #![cfg_attr(not(feature = "std"), no_std)]
 #![deny(unsafe_code)]
 #![deny(clippy::all)]